Make team stats module
